requirements.txt
Flask==3.0.0
gunicorn==21.2.0
flask-cors
psycopg2-binary
PyJWT==2.8.0
